feat: add unit tests for pinia stores
