+++ frontend/app/tests/unit/specs/components/helper/tab-navigation.spec.ts
