+++ frontend/app/tests/unit/specs/composables/accounts/blockie.spec.ts
