+++ frontend/app/tests/unit/specs/composables/defi.spec.ts
