+++ frontend/app/tests/unit/specs/data/amount_formatter.spec.ts
