+++ frontend/app/tests/unit/specs/store/settings/frontend.spec.ts
