+++ frontend/app/tests/unit/specs/store/settings/utils.spec.ts
